requirements.txt
pybit==5.13.0
python-dotenv==1.2.1
requests==2.32.5
